add TodoItem compoent and fix code mistakes
